+++ inc/functions.php
@@ -142,6 +142,7 @@ function caweb_template_colors() {
 function caweb_nav_menu_types() {
 	$menu_types = array(
 		'dropdown'     => 'Drop Down',
+		'megadropdown' => 'Mega Drop',
 		'singlelevel'  => 'Single Level',
 	);